backend/requirements.txt
bcrypt==4.1.2
blinker==1.7.0
click==8.1.7
colorama==0.4.6
Flask==3.0.0
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector==2.2.9
mysql-connector-python==8.2.0
protobuf==4.21.12
PyJWT==2.8.0
SQLAlchemy==2.0.24
typing_extensions==4.9.0
Werkzeug==3.0.1
